added SLabel #25
1. refactored SButton to be SLabel

2. made SLabel to support capitalization setting.
